+++ voip_sip_webrtc_twilio/models/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+from . import voip_twilio
+from . import voip_call
+from . import voip_call_wizard
+from . import res_partner
+from . import voip_number