+++ voip_sip_webrtc_twilio/models/voip_call.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+import logging
+_logger = logging.getLogger(__name__)
+from openerp import api, fields, models
+
+class VoipCallInheritTWilio(models.Model):
+
+    _inherit = "voip.call"
+
+    twilio_sid = fields.Char(string="Twilio SID")
+    twilio_account_id = fields.Many2one('voip.twilio', string="Twilio Account")
+    currency_id = fields.Many2one('res.currency', string="Currency")
+    price = fields.Float(string="Price")
+    margin = fields.Float(string="Margin")