+++ voip_sip_webrtc/doc/changelog.rst
@@ -1,10 +1,14 @@
+v1.0.3
+======
+* Remove borken SIP widget
+
 v1.0.2
 ======
 * Convert sip_register, sip_invite and rtp code to python 3
 
 v1.0.1
 ======
-* unhide voip account field in user form view
+* Unhide voip account field in user form view
 
 v1.0
 ====